renaming recent articles and episodes components
